+++ views/movies.ejs
@@ -1,7 +1,7 @@
-<h1>Article Search List</h1>
+<h1>Click a Poster to get More Info</h1>
 
 <% results.forEach(function(item) { %>
 <div>
-  <a href='/moreinfo/<%= item.imdbID %>'><img src='<%= item.Poster %>'></a>
+  <a href='/movies/moreinfo/<%= item.imdbID %>'><img src='<%= item.Poster %>'></a>
 </div>
 <% }); %>
