+++ views/search.ejs
@@ -1,6 +1,6 @@
-<h1>Search</h1>
+<h1>Search For Movies</h1>
 <form action="/movies" method="get">
-  <input type="text" name="search">
+  <input type="text" name="searchTerm">
   <input type="submit">
 </form>
